Struct tcod::console::Offscreen

pub struct Offscreen { /* fields omitted */ }

A type representing secondary consoles

Offscreen consoles allow you draw on secondary consoles as you would on Root consoles, and then blit their contents onto other consoles (including Root). There are some limitations compared to Root consoles, however:

Functions manipulating the main window or handling user input are limited to the Root console

Offscreen consoles may not be flushed to the screen directly

Examples

Creating an Offscreen console

use tcod::console::Offscreen;

let width = 80;
let height = 20;
let offscreen = Offscreen::new(width, height);

Blitting an Offscreen console to the Root console:

use tcod::console as console;
use tcod::console::{Root, Offscreen};

fn main() {
    let mut root = Root::initializer().init();

    let mut direct = Offscreen::new(20, 20);
    console::blit(&direct, (0, 0), (20, 20), &mut root, (0, 0), 1.0, 1.0);
}

See the documentation for blit for a detailed description of the function parameters and a more in-depth example.

Methods

impl Offscreen

fn new(width: i32, height: i32) -> Offscreen

Creates a new Offscreen console instance