Crate tcod

libtcod bindings for Rust

Description

libtcod a.k.a. "The Doryen Library" is a free, fast, portable and uncomplicated API for roguelike developpers providing lots of useful functions, such as:

Text-based graphics API

Colors

Keyboard and mouse input

Path finding

Field of View

Portability (works on Windows, Linux and OS X)

For the full set of features see the libtcod features page.

All raw bindings are available via the tcod-sys crate, however the tcod-rs library aims to provide safe, Rust-style wrappers for most of libtcod. These wrappers are not yet complete, however.

Features already implemented:

Colors

Console

Most of the System layer

Field of View

Map

Pathfinding

Line toolkit

Noise

BSP Toolkit

Features that are not planned to be implemented:

This are features that Rust already provides a good (and in most casese more idiomatic) solution for:

Filesystem utilities

Containers

Pseudorandom generators

Compression utilities

Reexports

pub use colors::Color;

pub use console::{Console, RootInitializer, BackgroundFlag, Renderer, FontLayout, FontType, TextAlignment};

pub use map::Map;

Modules

bsp
Port of BSP toolkit.

chars

colors

console
The console emulator handles the rendering of the game screen and the keyboard input

image

input

line
Port of line drawing toolkit.

map

namegen

noise
Noise toolkit.

pathfinding

random

system

Macros

tcod_print!

Traits

AsNative

FromNative

Type Definitions

OffscreenConsole

RootConsole