Struct tcod::bsp::Bsp

pub struct Bsp<'a> { /* fields omitted */ }

This struct encapsulates TCOD_bsp_t. It mirrors original's fields (x, y, etc.) with the exception of horizontal. See example.

Examples

let mut bsp = Bsp::new_with_size(0, 0, 50, 60);

assert_eq!(bsp.x, 0);
assert_eq!(bsp.y, 0);
assert_eq!(bsp.w, 50);
assert_eq!(bsp.h, 60);
assert_eq!(bsp.horizontal(), false);

bsp.x = 10;
bsp.y = 20;
bsp.set_horizontal(true);

Methods

impl<'a> Bsp<'a>

pub fn new_with_size(x: i32, y: i32, w: i32, h: i32) -> Self

pub fn remove_sons(&mut self)

pub fn split_once(&mut self, horizontal: bool, position: i32)

pub fn split_recursive(
&mut self,
randomizer: Option<Rng>,
nb: i32,
min_h_size: i32,
min_v_size: i32,
max_h_ratio: f32,
max_v_ratio: f32
)

pub fn resize(&mut self, x: i32, y: i32, w: i32, h: i32)

pub fn left(&self) -> Option<Self>

Returns Some(Bsp) with left subtree, or None if the BSP has not been split.

pub fn right(&self) -> Option<Self>

Returns Some(Bsp) with right subtree, or None if the BSP has not been split.

pub fn father(&self) -> Option<Self>

Returns Some(Bsp) with father, or None if the node is root.

pub fn is_leaf(&self) -> bool

pub fn contains(&self, cx: i32, cy: i32) -> bool

pub fn find_node(&self, cx: i32, cy: i32) -> Option<Self>

pub fn horizontal(&self) -> bool

pub fn set_horizontal(&mut self, h: bool)

pub fn traverse<F>(&self, order: TraverseOrder, callback: F) -> bool where
F: FnMut(&mut Bsp) -> bool,

Instead of 5 traverse* functions as in original API, Rust binding provides a single traverse function with an order parameter.

Examples

   let bsp = Bsp::new_with_size(0, 0, 50, 50);
   let mut counter = 0;

   bsp.traverse(TraverseOrder::PreOrder, |node| {
       counter += 1;
       true
   });
   assert_eq!(counter, 1);