Enum tcod::console::BackgroundFlag

pub enum BackgroundFlag {
    None,
    Set,
    Multiply,
    Lighten,
    Darken,
    Screen,
    ColorDodge,
    ColorBurn,
    Add,
    AddA,
    Burn,
    Overlay,
    Alph,
    Default,
}

This flag determines how a cell's existing background color will be modified by a new one

See libtcod's documentation for a detailed description of the different values.

Variants

None

Set

Multiply

Lighten

Darken

Screen

ColorDodge

ColorBurn

Add

AddA

Burn

Overlay

Alph

Default