Struct tcod::console::Root

pub struct Root {
    // some fields omitted
}

The console representing the main window of the application

This is the only console type capable of handling user input and flushing its contents onto the screen. There may only be one Root console at any given time, and it should be initialized at the start of the program.

Examples

Handling user input

tcod-rs provides two ways of handling user input: blocking or non-blocking. The following exaple will show the blocking method

use tcod::console::Root;
use tcod::input::Key::Special;
use tcod::input::KeyCode::{Up, Down, Left, Right};

fn main() {
    let mut root = Root::initializer().init();

    let keypress = root.wait_for_keypress(true);
    match keypress.key {
        Special(Up) => {}, // Handle arrow key up
        Special(Down) => {}, // Arrow key down
        Special(Left) => {},
        Special(Right) => {},
        _ => {}
    }
}

For a detailed description of possible values of keypress.key values see the Key and KeyCode enums.

Rendering

libtcod provides a wide variety of functions for changing how the console output looks like, including: changing the text and background colors, text alignment, etc. It also has several functions that are used to output text on consoles. For a complete list of both, see the Console trait's documentation. The basic structure of the rendering code:

use tcod::console::{Console, Root};

fn main() {
    let mut root = Root::initializer().init();

    root.clear();
    // Output style manipulation
    // Calling the output functions
    root.flush();
}

Methods

impl Root

fn initializer<'a>() -> RootInitializer<'a>

Returns an instance of a RootInitializer object, which can be used to customize the initialization of the Root console. Note that only RootInitializer::init will return the actual Root console instance. For a full list of initialization options, see the RootInitializer documentation.

Examples

use tcod::console::Root;

let mut root = Root::initializer()
    .size(80, 20)
    .title("Example")
    .fullscreen(true)
    .init();

fn is_fullscreen(&self) -> bool

Returns with true when the Root console is in fullscreen mode.

fn set_fullscreen(&mut self, fullscreen: bool)

Toggles between windowed and fullscreen mode.

fn set_keyboard_repeat(&mut self, initial_delay: i32, interval: i32)

This function changes the keyboard repeat times. The initial delay determines the number of milliseconds between the keypress and the time the keyboard repeat begins. The interval sets the time between the keyboard repeat events. With an initial delay of 0, the keyboard repeat feature is completely disabled.

fn disable_keyboard_repeat(&mut self)

Disables the keyboard repeat feature. Equivalent to set_keyboard_repeat with an initial delay of 0.

fn is_active(&self) -> bool

Returns true if the Root console is currently active.

fn get_fade(&self) -> u8

Returns the current fade amount (previously set by set_fade).

fn get_fading_color(&self) -> Color

Returns the current fade color (previously set by set_fade).

fn set_fade(&mut self, fade: u8, fading_color: Color)

This function defines the fading parameters, allowing to easily fade the game screen to/from a color. Once they are defined, the fading parameters are valid for ever. You don't have to call setFade for each rendered frame (unless you change the fading parameters).

fn wait_for_keypress(&mut self, flush: bool) -> KeyState

This function will wait for a keypress event from the user, returning the KeyState that represents the event. If flush is true, all pending keypresses are flushed from the keyboard buffer. If false, it returns the first element from it.

fn check_for_keypress(&self, status: KeyPressFlags) -> Option<KeyState>

This function checks if the user pressed a key. It returns the KeyState representing the event if they have, or None if they have not.

fn window_closed(&self) -> bool

Returns with true if the Root console has been closed.

fn flush(&mut self)

Flushes the contents of the Root console onto the screen.

fn set_window_title(&mut self, title: &str)

Sets the main window's title to the string specified in the argument.