Struct tcod::console::Root

pub struct Root { /* fields omitted */ }

Methods

impl Root

fn initializer<'a>() -> RootInitializer<'a>

Returns an instance of a RootInitializer object, which can be used to customize the initialization of the Root console. Note that only RootInitializer::init will return the actual Root console instance. For a full list of initialization options, see the RootInitializer documentation.

Examples

use tcod::console::Root;

let mut root = Root::initializer()
    .size(80, 20)
    .title("Example")
    .fullscreen(true)
    .init();

fn is_fullscreen(&self) -> bool

Returns with true when the Root console is in fullscreen mode.

fn set_fullscreen(&mut self, fullscreen: bool)

Toggles between windowed and fullscreen mode.

fn set_keyboard_repeat(&mut self, initial_delay: i32, interval: i32)

This function changes the keyboard repeat times. The initial delay determines the number of milliseconds between the keypress and the time the keyboard repeat begins. The interval sets the time between the keyboard repeat events. With an initial delay of 0, the keyboard repeat feature is completely disabled.

fn disable_keyboard_repeat(&mut self)

Disables the keyboard repeat feature. Equivalent to set_keyboard_repeat with an initial delay of 0.

fn is_active(&self) -> bool

Returns true if the Root console is currently active.

fn has_focus(&self) -> bool

Returns true if the Root console has focus.

fn get_fade(&self) -> u8

Returns the current fade amount (previously set by set_fade).

fn get_fading_color(&self) -> Color

Returns the current fade color (previously set by set_fade).

fn set_fade(&mut self, fade: u8, fading_color: Color)

This function defines the fading parameters, allowing to easily fade the game screen to/from a color. Once they are defined, the fading parameters are valid for ever. You don't have to call setFade for each rendered frame (unless you change the fading parameters).

fn wait_for_keypress(&mut self, flush: bool) -> Key

This function will wait for a keypress event from the user, returning the KeyState that represents the event. If flush is true, all pending keypresses are flushed from the keyboard buffer. If false, it returns the first element from it.

fn check_for_keypress(&self, status: KeyPressFlags) -> Option<Key>

This function checks if the user pressed a key. It returns the KeyState representing the event if they have, or None if they have not.

fn window_closed(&self) -> bool

Returns with true if the Root console has been closed.

fn flush(&mut self)

Flushes the contents of the Root console onto the screen.

fn set_window_title<T>(&mut self, title: T) where
T: AsRef<str>,

Sets the main window's title to the string specified in the argument.

fn render_credits(&self, x: i32, y: i32, alpha: bool) -> bool

Embeds libtcod credits in a console. Returns true when the credits screen is finished.

fn map_ascii_code_to_font(
&mut self,
ascii_code: i32,
font_char_x: i32,
font_char_y: i32
)

Maps a single ASCII code to a character in a bitmap font.

Arguments

ascii_code: The ASCII code to map

font_char_x/font_char_y: The coordinate of the character in the bitmap font (in characters, not pixels)

fn map_ascii_codes_to_font(
&mut self,
ascii_code: i32,
nb_codes: i32,
font_char_x: i32,
font_char_y: i32
)

Maps consecutive ASCII codes to consecutive characters in a bitmap font.

Arguments

ascii_code: The first ASCII code to map

'nb_codes`: Number of conescutive ASCII codes to map

font_char_x/font_char_y: The coordinate of the character in the bitmap font (in characters, not pixels) corresponding to the first ASCII code

fn map_string_to_font(&mut self, s: &str, font_char_x: i32, font_char_y: i32)

Maps ASCII codes from a string to consecutive characters in a bitmap font.

Arguments

s: String containing ASCII codes to map

font_char_x/font_char_y: The coordinate of the character in the bitmap font (in characters, not pixels) corresponding to the first ASCII code in the string